+++ Microsoft.Azure.Cosmos/src/FeedRange/FeedRanges/FeedRangePartitionKeyRange.cs
@@ -8,16 +8,12 @@ namespace Microsoft.Azure.Cosmos
     using System.Collections.Generic;
     using System.Threading;
     using System.Threading.Tasks;
-    using Microsoft.Azure.Cosmos.Core.Trace;
     using Microsoft.Azure.Cosmos.Routing;
-    using Newtonsoft.Json;
-    using Newtonsoft.Json.Linq;
 
     /// <summary>
     /// FeedRange that represents a Partition Key Range.
     /// Backward compatibility implementation to transition from V2 SDK queries that were filtering by PKRangeId.
     /// </summary>
-    [JsonConverter(typeof(FeedRangePartitionKeyRangeConverter))]
     internal sealed class FeedRangePartitionKeyRange : FeedRangeInternal
     {
         public string PartitionKeyRangeId { get; }
@@ -26,8 +22,6 @@ namespace Microsoft.Azure.Cosmos
         {
             this.PartitionKeyRangeId = partitionKeyRangeId;
         }
-
-        public override string ToJsonString() => JsonConvert.SerializeObject(this);
 
         public override async Task<List<Documents.Routing.Range<string>>> GetEffectiveRangesAsync(
             IRoutingMapProvider routingMapProvider,
@@ -59,40 +53,5 @@ namespace Microsoft.Azure.Cosmos
         }
 
         public override string ToString() => this.PartitionKeyRangeId;
-
-        public static new bool TryParse(
-            JObject jObject,
-            JsonSerializer serializer,
-            out FeedRangeInternal feedRangeInternal)
-        {
-            try
-            {
-                feedRangeInternal = FeedRangePartitionKeyRangeConverter.ReadJObject(jObject, serializer);
-                return true;
-            }
-            catch (JsonReaderException)
-            {
-                DefaultTrace.TraceError("Unable to parse FeedRange for PartitionKeyRange");
-                feedRangeInternal = null;
-                return false;
-            }
-        }
-
-        public static new bool TryParse(
-            string jsonString,
-            out FeedRangeInternal feedRangeInternal)
-        {
-            try
-            {
-                feedRangeInternal = JsonConvert.DeserializeObject<FeedRangePartitionKeyRange>(jsonString);
-                return true;
-            }
-            catch (JsonReaderException)
-            {
-                DefaultTrace.TraceError("Unable to parse FeedRange for PartitionKeyRange");
-                feedRangeInternal = null;
-                return false;
-            }
-        }
     }
 }
